+++ market-access/translations.js
@@ -1,0 +1,15 @@
+module.exports= {
+    'quantity_unshield_maize': {
+        'en': 'Quantity of Unshelled Maize ( Use big bags of 100Kgs)',
+        'ki': 'Ingono y\'ibigori bidahunguye (Mu mifuka y\'ibiro 100)'
+    },
+    'maize_availability': {
+        'en': 'When  Maize will be available',
+        'ki': 'Ni ryari ibigori bizaba byumye neza'
+    },
+    'payment_advance': {
+        'en': 'Do you want an Advance payment for your maize? \n1)Yes\n2)No',
+        'ki': 'Ese ukeye avansi\n1) Yego\n2) Oya'
+    }
+
+};